Create Area and Perimeter of Triangle. 1
Vihaan wants to know how to find the area and perimeter of the triangle.could you please the help him to find it and implements in programming. ** 1. Area of triangle:-1/2*b*h 2. Perimeter of triangle:- sum of threeside.**

Input Format

input consist of 5 integer First and second input corresponds to base and height. Third , Fourth and fifth are side of the triangle

Constraints

No Constraints

Output Format

print the area and perimeter of the triangle.

Sample Input 0

5
6
1
2
3
Sample Output 0

Area of Triangle is 15.00
Perimeter of Triangle is 6.00
Sample Input 1

12
6
5
12
13
Sample Output 1

Area of Triangle is 36.00
Perimeter of Triangle is 30.00