Create Sum of given Number 1
Write a program to print the sum of the given numbers. The values must be scanned until the user enters -1 as value. Note: Use a do-while loop.

Input Format

Input consists of a list of integers.

Constraints

No Constraints

Output Format

The output consists of the sum of the given integers.

Sample Input 0

1
2
3
4
-1
Sample Output 0

10
Sample Input 1

-1
Sample Output 1

0