Create Account Balance 3
Ramu after receiving his first salary decided to go to the bank to deposit his savings in his account. Before that, he does not know the process that is to be undertaken in the bank. Help him in understanding the process of depositing money and withdrawal of money. Use Switchcase to implement this concept

Input Format

First input corresponds to the account balance. Second input corresponds to the process to be carried out.

Constraints

No Constraints

Output Format

If the input is 1, the amount should be deposited. If the input is 2, the amount should be withdrawn and for other inputs display “Invalid Input”. If the amount to be withdrawn is greater than the balance, display “Insufficient Balance”.

Sample Input 0

1000
3
Sample Output 0

Invalid Input
Sample Input 1

2000
1
500
Sample Output 1

2500
Sample Input 2

200
2
2000
Sample Output 2

Insufficient Balance
Contest ends in 8 days
Submissions: 128
Max Score: 10
Difficulty: Medium
Rate This Challenge:

    
More