Create Search an Element 7
Dhiya wants to search for the presence of an element in an array.could you please help her to search it and implements in program.

Input Format

Input consists of one integer and list of array elements. First input corresponds to the array size. Second input corresponds to the array elements. Third input corresponds to finding an array element.

Constraints

No constraints

Output Format

Print the Statement whether the given elements is Presented in the array or not.

Sample Input 0

6
1
2
3
4
5
6
7
Sample Output 0

7 is not presented in an array.
Sample Input 1

5
2
5
3
6
2
3
Sample Output 1

3 is presented in an array.