Create Fibonacci series 144
Write the program to calculate the sum of the Fibonacci series between the range of the two values. The n is the start of the range value and m is the end of the range value. Finally, return the sum of the range value of the Fibonacci Series.

Input Format

Input consists of two integers.

Constraints

N is greater than equal to one and lesser than equal to 20

M is greater than equal to one and lesser than equal to 20

Output Format

The sum of the range value.

If the constraints are more than the input print the statement is "Invalid Input".

Sample Input 0

3
5
Sample Output 0

The Sum of Fibonacci value is 6.0
Sample Input 1

22
23
Sample Output 1

Invalid Input