Create Count distinct elements 8
Write a program to count the number of distinct elements in an array.

Input Format

Input consists of 1 integer and 1 array. The first integer corresponds to the size of the array.

Constraints

No Constraints

Output Format

Print the no of the distinct element

Sample Input 0

6
2
6
6
4
5
5
Sample Output 0

There are 4 distinct elements in the array.
Sample Input 1

4
1
1
1
1
Sample Output 1

There are 1 distinct element in the array.