Create Difference of the character 1
Given a string S(input consisting) of '' and '#'. The length of the string is variable. The task is to find the minimum number of '' or '#' to make it a valid string. The string is considered valid if the number of '' and '#' are equal. The '' and '#' can be at any position in the string.

Note : The output will be a positive or negative integer based on number of '*' and '#' in the input string.

(>#): positive integer (#>): negative integer (#=*): 0 Example 1:

Input 1:

* -> Value of S
Output :

0 → number of * and # are equal

Input Format

Input consist of one string

Constraints

No constraints

Output Format

Execute the given output format.check the output data whether it is odd or even.

Sample Input 0

#**#**
Sample Output 0

The Difference of the character in the given string: 02
Sample Input 1

**##**#
Sample Output 1

The Difference of the character in the given string: 001
Sample Input 2

##*#**
Sample Output 2

0