Create Oops-Class-Object 1
Write a program to create a class called "Rectangle" with width and height attributes. Calculate the area and perimeter of the rectangle.

Input Format

Input Consists of two integer. First one is for width and second one is for height.

Constraints

Both value should be lesser than equals to 10

Output Format

Execute the values of both are and perimeter of the rectangle.

Sample Input 0

5
4
Sample Output 0

Here your Answers!!!
Area of the Rectangle is 20.00
Perimeter of the Rectangle is 18.00
Sample Input 1

12
4
Sample Output 1

Hello User,Kindly Enter the valid Input.